Fix compile errors

+++ pong_core/src/constants.rs
@@ -1,6 +1,6 @@
-pub const LCD_HEIGHT: usize = 272;
-pub const LCD_WIDTH: usize = 480;
-pub const PADDLE_OFFSET: usize = 50;
-pub const PADDLE_SPEED: usize = 1;
-pub const PADDLE_HEIGHT: usize = 90;
-pub const PADDLE_WIDTH: usize = 10;
+pub const LCD_HEIGHT: i32 = 272;
+pub const LCD_WIDTH: i32 = 480;
+pub const PADDLE_OFFSET: i32 = 50;
+pub const PADDLE_SPEED: i32 = 1;
+pub const PADDLE_HEIGHT: i32 = 90;
+pub const PADDLE_WIDTH: i32 = 10;
